+++ lib/heroicons/icon.ex
@@ -14,7 +14,7 @@ defmodule Heroicons.Icon do
 
     * `:name` - the name of the icon
 
-    * `:file` - the binary of the icon
+    * `:file` - the binary content of the file
 
   """
   defstruct [:type, :name, :file]
@@ -29,36 +29,22 @@ defmodule Heroicons.Icon do
       |> Path.split()
       |> Enum.take(-3)
       |> case do
+        ["16", "solid", name] -> ["micro", name]
         ["20", "solid", name] -> ["mini", name]
         ["24", "solid", name] -> ["solid", name]
         ["24", "outline", name] -> ["outline", name]
       end
 
     name = Path.rootname(name)
-    file = File.read!(filename)
 
-    struct!(__MODULE__, type: type, name: name, file: file)
-  end
+    file =
+      filename
+      |> File.read!()
+      |> String.split("\n")
+      |> Enum.drop(1)
+      |> Enum.drop(-2)
+      |> Enum.map(&String.trim/1)
 
-  @doc "Converts opts to HTML attributes"
-  @spec opts_to_attrs(keyword) :: list
-  def opts_to_attrs(opts) do
-    for {key, value} <- opts do
-      key =
-        key
-        |> Atom.to_string()
-        |> String.replace("_", "-")
-        |> Phoenix.HTML.Safe.to_iodata()
-
-      value = Phoenix.HTML.Safe.to_iodata(value)
-
-      [?\s, key, ?=, ?", value, ?"]
-    end
-  end
-
-  @doc "Inserts HTML attributes into an SVG icon"
-  @spec insert_attrs(binary, keyword) :: Phoenix.HTML.safe()
-  def insert_attrs("<svg" <> rest, attrs) do
-    Phoenix.HTML.raw(["<svg", attrs, rest])
+    %__MODULE__{type: type, name: name, file: file}
   end
 end